+++ data/datasets/change_detection_datasets/bi_temporal/urb3dcd_dataset.py
@@ -414,14 +414,14 @@ class Urb3DCDDataset(Base3DCDDataset):
         print("Loading " + files['pc_2_filepath'])
         nameInPly = self.VERSION_MAP[self.version]['nameInPly']
 
-        # Load first point cloud (only has XYZ coordinates)
-        pc1_data = utils.io.load_point_cloud(files['pc_1_filepath'], nameInPly=nameInPly, name_feat="label_ch")
+        # Load first point cloud (only has XYZ coordinates) with float64 precision
+        pc1_data = utils.io.load_point_cloud(files['pc_1_filepath'], nameInPly=nameInPly, name_feat="label_ch", dtype=torch.float64)
         pc1_xyz = pc1_data['pos']  # Extract position from dictionary
         # Add ones feature
         pc1_features = torch.ones((pc1_xyz.size(0), 1), dtype=pc1_xyz.dtype)  # [N, 1]
 
-        # Load second point cloud (XYZ coordinates + label)
-        pc2_data = utils.io.load_point_cloud(files['pc_2_filepath'], nameInPly=nameInPly, name_feat="label_ch")
+        # Load second point cloud (XYZ coordinates + label) with float64 precision
+        pc2_data = utils.io.load_point_cloud(files['pc_2_filepath'], nameInPly=nameInPly, name_feat="label_ch", dtype=torch.float64)
         pc2_xyz = pc2_data['pos']  # Extract position from dictionary
         # Add ones feature
         pc2_features = torch.ones((pc2_xyz.size(0), 1), dtype=pc2_xyz.dtype)  # [N, 1]
